requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
google-generativeai==0.8.3
google-genai>=1.0.0
nbformat>=5.9.0
python-dotenv==1.0.1
requests==2.32.3
aiofiles==24.1.0
markdown2==2.5.0
Pillow==10.4.0
PyMuPDF>=1.24.0
httpx==0.27.2
# PaperBanana dependencies (vendor clone at vendor/paperbanana/)
pydantic>=2.0
pydantic-settings>=2.0
pyyaml>=6.0
typer>=0.12
rich>=13.0
matplotlib>=3.8
pandas>=2.0
tenacity>=8.0
structlog>=24.0